check combine cards

+++ Event/interface/JetCollection.h
@@ -39,7 +39,7 @@ class JetCollection : public PhysicsObjectCollection< Jet > {
 	JetCollection HEM1516DownCollection() const;
 
         //add flavor split JES variations
-        JetCollection JECGroupedFlavorQCDCollection(unsigned flavor,std::string source, bool up) const;
+        JetCollection JECGroupedFlavorQCDCollection(unsigned long flavor,std::string source, bool up) const;
         //JetCollection JECGroupedFlavorQCDDownCollection()(unsigned flavor) const;
         //JetCollection JECGroupedFlavorQCDUpCollection()(unsigned flavor) const;
 
@@ -85,8 +85,8 @@ class JetCollection : public PhysicsObjectCollection< Jet > {
 	JetCollection buildVariedCollection( Jet (Jet::*variedJet)(std::string) const, std::string ) const;
 
         //add flavor split JES variations
-        JetCollection buildVariedCollection_FlavorSet( Jet (Jet::*variedJet)(const unsigned) const, unsigned, unsigned ) const;
-        JetCollection buildVariedCollection_FlavorSet( Jet (Jet::*variedJet)(std::string) const, std::string, unsigned ) const;
+        JetCollection buildVariedCollection_FlavorSet( Jet (Jet::*variedJet)(const unsigned) const, unsigned, unsigned long) const;
+        JetCollection buildVariedCollection_FlavorSet( Jet (Jet::*variedJet)(std::string) const, std::string, unsigned long) const;
 
         //number of b-taged jets with variation
         std::vector< size_type > countsAnyVariation( bool ( Jet::*passSelection )() const ) const;
